+++ Mecanillama.API/Customers/Domain/Repositories/ICustomerRepository.cs
@@ -5,7 +5,7 @@ namespace Mecanillama.API.Customers.Domain.Repositories;
 public interface ICustomerRepository {
     Task<IEnumerable<Customer>> ListAsync();
     Task AddAsync(Customer customer);
-    Task<Customer> FindByIdAsync(long id);
+    Task<Customer> FindByIdAsync(int id);
     Customer FindById(int id);
     Task<Customer> FindByEmailAsync(string email);
     public bool ExistsByEmail(string email);
